Allow `OrganisationFeedContent` to be used for other content
`OrganisationFeedContent` currently takes in a organisation name. By
changing that to accept a part of a rummager search query (like
`filter_organisations=organisation_name`) we can make the class
re-usable for other types of content.
